move all app scripts to src directory
