+++ Dockerfile
@@ -8,8 +8,4 @@ RUN pip install --no-cache-dir --upgrade -r /code/requirements.txt
 
 COPY . /code/
 
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "80"]
-
-
-
-
+CMD ["uvicorn", "src.main:app", "--host", "0.0.0.0", "--port", "80"]
